Refactor: コードにコメントを追加 + 余分な空白削除
