+++ back/src/utils/errorResponse.ts
@@ -2,6 +2,7 @@ const _createObjectWithError = (text: string) => {
   return { error: text }
 }
 
+// APIから返すエラーメッセージ一覧
 export const ACCESS_TOKEN_EXPIRE_ERROR = _createObjectWithError('AccessTokenExpired')
 
 export const INVALID_TOKEN_ERROR = _createObjectWithError('InvalidToken')
