+++ back/src/utils/jwt.ts
@@ -2,8 +2,13 @@ import jwt from 'jsonwebtoken'
 import { jwtPayload } from '../types/auth';
 import { Response } from 'express';
 
-const SECRET_KEY: string = process.env.JWT_SECRET_KEY || 'flabkdfafkadkvjkavv3d23adkjv'
-const COOKIE_NAME: string = 'jwt_token'
+const SECRET_KEY = process.env.JWT_SECRET_KEY
+if (!SECRET_KEY) {
+  console.log('変数SECRET_KEYが定義されていません');
+  process.exit(1);
+}
+const COOKIE_NAME = 'jwt_token'
+const ACCESS_TOKEN_EXPIRESIN = Number(process.env.ACCESS_TOKEN_EXPIRESIN) || (30 * 60)
 
 export const decodeJwt = (token: string) => {
   const decoded = jwt.decode(token, { json: true });
@@ -11,7 +16,8 @@ export const decodeJwt = (token: string) => {
 }
 
 export const createJwt = (payload: jwtPayload): string => {
-  const token: string = jwt.sign(payload, SECRET_KEY, { expiresIn: "30m" });
+  // デフォルトの暗号方式はHS256（共通鍵）
+  const token: string = jwt.sign(payload, SECRET_KEY, { expiresIn: ACCESS_TOKEN_EXPIRESIN });
   return token;
 }
