+++ front/src/features/records/recordsSlice.ts
@@ -4,7 +4,7 @@ import type { PostRecordParams, RecordsState, TermsRecords } from "./types/recor
 import type { AppDispatch, RootState } from "../../reduxStore/store";
 import { devLog } from "../../utils/logDev";
 import { getAxiosErrorMessageFromStatusCode } from "../../utils/errorHandler/axiosError";
-import { fetchWithTokenRefresh } from "../../utils/asyncFetch/fetchWithTokenRefresh";
+import { fetchWithTokenRefresh } from "../../utils/fetch/fetchWithTokenRefresh";
 import { INVALID_REFRESH_TOKEN } from "../../utils/apiErrors/errorMessages";
 import { resetStateOfUser } from "../session/slices/sessionSlice";
