+++ front/src/features/todos/todoSlice.ts
@@ -5,7 +5,7 @@ import { defaultTodos } from './defaultTodos';
 import { devLog } from '../../utils/logDev';
 import { getAxiosErrorMessageFromStatusCode } from '../../utils/errorHandler/axiosError';
 import { sortTodosByDeadline } from './utils/todosSort';
-import { fetchWithTokenRefresh } from '../../utils/asyncFetch/fetchWithTokenRefresh';
+import { fetchWithTokenRefresh } from '../../utils/fetch/fetchWithTokenRefresh';
 import { INVALID_REFRESH_TOKEN } from '../../utils/apiErrors/errorMessages';
 import { resetStateOfUser } from '../session/slices/sessionSlice';
