+++ front/src/utils/apiErrors/errorMessages.ts
@@ -1,3 +1,4 @@
+// APIから返される特徴的なエラーメッセージ（特有の処理が必要なもの）
 export const ACCESS_TOKEN_EXPIRE_ERROR = 'AccessTokenExpired'
 
 export const INVALID_TOKEN_ERROR = 'InvalidToken'
