+++ front/src/utils/fetch/auth.ts
@@ -1,7 +1,7 @@
-import { clientCredentials } from "./axios"
-import { devLog } from "./logDev";
+import { clientCredentials } from "../axios"
+import { devLog } from "../logDev";
 
-// 認証トークンを送信し、ログイン状態を確認する
+// 認証トークン検証エンドポイントにfetchし、ログイン状態を確認
 export const checkAuth = async() => {
   try {
     const res = await clientCredentials.get('/auth/check');
