+++ front/src/utils/logDev.ts
@@ -1,3 +1,4 @@
+// 開発環境でのみログを表示
 export const devLog = (...args: unknown[]) => {
   if (!import.meta.env.DEV) return;
   console.log(...args);
