+++ front/src/utils/modelCtl.ts
@@ -1,5 +1,6 @@
 type dialogOrNull = HTMLDialogElement | null;
 
+// Modal操作
 export function openModal(id: string) {
   const dialogHtml = document.getElementById(id) as dialogOrNull;
   dialogHtml?.showModal();
